Some minor phonological rule organization and cleanup
Organized the rules into groups, and cleaned up some notes left in for
things that still need to be done. Fixed a couple rules which needed to
be changed.

Next step is ordering rules, then adding new ones in. Now that the rules
which produce the appropriate sound system are in, the next set of rules
can be freer and more fun! :)